fixes #10271 - ensure a host and image are created on OpenStack

+++ app/services/foreman_snapshot/snapshot.rb
@@ -49,7 +49,7 @@ module ForemanSnapshot
         )
       end
     ensure
-      template.unlink
+      template.unlink if template
     end
 
     private
